+++ integration_test/screenshot_test.dart
@@ -474,12 +474,6 @@ class _FakeFirstLogPrompt extends FirstLogPromptNotifier {
 
   @override
   Future<void> markShown() async {}
-
-  @override
-  void show() {}
-
-  @override
-  Future<void> dismiss() async {}
 }
 
 class _FakeWeatherOptIn extends WeatherOptInNotifier {